moves results card component
