+++ src/components/ResultsSidebar/ResultsCard/OpenNowCheck.ts
